Tracker hits were previously handled using the SVT tracker hits driver, but this is not strictly accurate. The SVT driver outputs FPGA data, which is not appropriate for calorimeter scoring plane hits or hodoscope hits. Instead, the shared functionality has been removed to a separate SimTrackerHitReadoutDriver for general use. This driver is then extended by the new SVTTrackerHitReadoutDriver to include FPGA support.
