+++ tracking/src/main/java/org/hps/readout/svt/SVTTrackerHitReadoutDriver.java
@@ -5,27 +5,19 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
-import org.hps.readout.ReadoutDataManager;
-import org.hps.readout.SLICDataReadoutDriver;
+import org.hps.readout.SimTrackerHitReadoutDriver;
 import org.hps.readout.util.collection.LCIOCollection;
 import org.hps.readout.util.collection.LCIOCollectionFactory;
 import org.hps.readout.util.collection.TriggeredLCIOData;
 import org.lcsim.detector.tracker.silicon.HpsSiSensor;
 import org.lcsim.detector.tracker.silicon.HpsTestRunSiSensor;
-import org.lcsim.event.MCParticle;
-import org.lcsim.event.SimTrackerHit;
 import org.lcsim.geometry.Detector;
 import org.lcsim.geometry.Subdetector;
 
-public class SimTrackerHitReadoutDriver extends SLICDataReadoutDriver<SimTrackerHit> {
+public class SVTTrackerHitReadoutDriver extends SimTrackerHitReadoutDriver {
     private Subdetector detector = null;
     private LCIOCollection<FpgaData> fpgaDataParams = null;
-    
-    public SimTrackerHitReadoutDriver() {
-        super(SimTrackerHit.class, 0xc0000000);
-    }
     
     @Override
     public void detectorChanged(Detector detector) {
@@ -43,7 +35,15 @@ public class SimTrackerHitReadoutDriver extends SLICDataReadoutDriver<SimTracker
         fpgaDataParams = LCIOCollectionFactory.produceLCIOCollection(FpgaData.class);
     }
     
+    @Override
     protected Collection<TriggeredLCIOData<?>> getOnTriggerData(double triggerTime) {
+        // The base class already provides outputs the truth hits
+        // associated with the output hits. The SVT also requires
+        // that FPGA data be written to the output file. Get the
+        // truth data from the superclass, so that it may be merged
+        // with the FPGA data.
+        Collection<TriggeredLCIOData<?>> superData = super.getOnTriggerData(triggerTime);
+        
         // Get the FPGA data.
         List<FpgaData> fpgaData = new ArrayList<FpgaData>(makeFPGAData(detector).values());
         
@@ -51,33 +51,15 @@ public class SimTrackerHitReadoutDriver extends SLICDataReadoutDriver<SimTracker
         TriggeredLCIOData<FpgaData> fpgaCollection = new TriggeredLCIOData<FpgaData>(fpgaDataParams);
         fpgaCollection.getData().addAll(fpgaData);
         
-        // Get the truth hits in the indicated time range.
-        Collection<SimTrackerHit> truthHits = ReadoutDataManager.getData(triggerTime - getReadoutWindowBefore(), triggerTime + getReadoutWindowAfter(), collectionName, SimTrackerHit.class);
-        
-        // MC particles need to be extracted from the truth hits
-        // and included in the readout data to ensure that the
-        // full truth chain is available.
-        Set<MCParticle> truthParticles = new java.util.HashSet<MCParticle>();
-        for(SimTrackerHit simHit : truthHits) {
-            ReadoutDataManager.addParticleParents(simHit.getMCParticle(), truthParticles);
-        }
-        
-        // Create the truth MC particle collection.
-        LCIOCollectionFactory.setCollectionName("MCParticle");
-        LCIOCollectionFactory.setProductionDriver(this);
-        LCIOCollection<MCParticle> truthParticleCollection = LCIOCollectionFactory.produceLCIOCollection(MCParticle.class);
-        TriggeredLCIOData<MCParticle> truthParticleData = new TriggeredLCIOData<MCParticle>(truthParticleCollection);
-        truthParticleData.getData().addAll(truthParticles);
-        
-        // Create a general list for the collection.
-        List<TriggeredLCIOData<?>> collectionsList = new ArrayList<TriggeredLCIOData<?>>(2);
+        // Add the FPGA data to the output data from the superclass.
+        List<TriggeredLCIOData<?>> collectionsList = new ArrayList<TriggeredLCIOData<?>>((superData == null ? 0 : superData.size()) + 1);
         collectionsList.add(fpgaCollection);
-        if(isPersistent()) { collectionsList.add(truthParticleData); }
         
         // Return the collections list result.
         return collectionsList;
     }
     
+    // TODO: SVT group should describe what this does.
     private Map<Integer, FpgaData> makeFPGAData(Subdetector subdetector) {
         double[] temps = new double[HPSSVTConstants.TOTAL_HYBRIDS_PER_FPGA * HPSSVTConstants.TOTAL_TEMPS_PER_HYBRID];
         for(int i = 0; i < HPSSVTConstants.TOTAL_HYBRIDS_PER_FPGA * HPSSVTConstants.TOTAL_TEMPS_PER_HYBRID; i++) {
@@ -99,10 +81,5 @@ public class SimTrackerHitReadoutDriver extends SLICDataReadoutDriver<SimTracker
         }
         
         return fpgaData;
-    }
-    
-    @Override
-    protected double getTimeNeededForLocalOutput() {
-        return isPersistent() ? getReadoutWindowAfter() : 0;
     }
 }
